fix: improve catalog views
