--- libs/catalog/src/components/CatalogFavorites/CatalogFavorites.module.scss
@@ -1,47 +1,0 @@
-.section {
-  overflow-anchor: none;
-  background:
-    linear-gradient(
-      180deg,
-      var(--cat-favorites-bg-start, var(--bg-accent-tertiary-alpha, #a972ff2e))
-        0%,
-      var(--cat-favorites-bg-end, var(--bg-accent-primary-alpha, #7da4ff26))
-        100%
-    ),
-    var(--cat-favorites-bg-base, var(--bg-layer-1, #0c101d));
-  border-bottom-color: var(
-    --cat-favorites-border,
-    var(--stroke-secondary, #242c42)
-  );
-}
-
-.title {
-  color: var(--cat-favorites-title-text, var(--text-primary, #eef1f7));
-}
-
-.count {
-  color: var(--cat-favorites-count-text, var(--text-secondary, #9fa6bd));
-}
-
-.gridPage > * {
-  animation: favFadeIn 180ms ease-out;
-}
-
-.paginationRow {
-  // Tailwind Preflight resets button{background-color:transparent} in @layer base,
-  // which overrides the DS's .bg-secondary in @layer ui (lower cascade priority).
-  // This unlayered rule wins over both named layers and restores the dot color.
-  :global(nav[aria-label='Pagination'] button.bg-secondary) {
-    background-color: var(--text-secondary, #9fa6bd);
-  }
-}
-
-@keyframes favFadeIn {
-  from {
-    opacity: 0;
-  }
-
-  to {
-    opacity: 1;
-  }
-}
